+++ CHANGELOG.rst
@@ -181,6 +181,8 @@ Changed
   `#326 <https://github.com/openego/eGon-data/issues/326>`_
 * Migrate osmTGmod to datasets
   `#305 <https://github.com/openego/eGon-data/issues/305>`_
+* Filter osm landuse areas, rename industrial sites tables and update load curve function
+  `#378 <https://github.com/openego/eGon-data/issues/378>`_
 * Remove version columns from eTraGo tables and related code
   `#384 <https://github.com/openego/eGon-data/issues/384>`_